tools/requirements.txt
arrow==0.17.0
OpenCC==1.1.3
prettytable==prettytable==3.10.0
PyYAML==5.4.1
requests==2.27.1
requirements-parser==0.5.0
simplejson==3.17.6
six==1.16.0
